Refactored a bit to reuse the MatrixPool for paths.
