+++ packages/flame/lib/extensions.dart
@@ -7,6 +7,7 @@ export 'src/extensions/list.dart';
 export 'src/extensions/matrix4.dart';
 export 'src/extensions/offset.dart';
 export 'src/extensions/paint.dart';
+export 'src/extensions/path.dart';
 export 'src/extensions/picture.dart';
 export 'src/extensions/rect.dart';
 export 'src/extensions/rectangle.dart';
